// src/app/models/staffType.ts
export enum StaffType{
    NURSE = "NURSE",
    TCAE = "TCAE",
    FISIO = "FISIO",
    SPEECH_THERAPIST = "SPEECH_THERAPIST",
    OCCUPATIONAL_THERAPY = "OCCUPATIONAL_THERAPY",
    PODIATRIST = "PODIATRIST",
    TEAP = "TEAP",
    TEDN = "TEDN",
    TEDS = "TEDS",
    TEL = "TEL",
    TEMN = "TEMN",
    TERD = "TERD",
    TER = "TER",
    TF = "TF",
    TSESPC = "TSESPC",
    TSSA = "TSSA",
    TSPRLPA = "TSPRLPA",
    TSPRLHI = "TSPRLHI",
    TSPRLST = "TSPRLST",
    SOCIAL_WORKER = "SOCIAL_WORKER",
    ADM = "ADM",
    ADMINISTRATIVE_ASSISTANT = "ADMINISTRATIVE_ASSISTANT",
    CC = "CC",
    CCUL = "CCUL",
    CCCTS = "CCCTS",
    CCAA = "CCAA",
    COOK = "COOK",
    POC = "POC",
    MONITOR = "MONITOR",
    POP = "POP",
    TEI = "TEI",
    TSA = "TSA",
    PHONE = "PHONE",
    TEMEII = "TEMEII",
    TEMEITF = "TEMEITF",
    TEMEISEA = "TEMEISEA",
    TMAC = "TMAC",
    TMMM = "TMMM",
    TMOA = "TMOA",
    TEE = "TEE",
    TIPRL = "TIPRL",
    WARDEN = "WARDEN",
    CLEANER = "CLEANER",
    LAUNDRY_STAFF = "LAUNDRY_STAFF",
    LABOURER = "LABOURER",
    HELPER = "HELPER",
}